Renamed script and settings table...
Now that the script is also creating a table in addition to the stored
procedures, the name “CREATE-SP” just didn’t fit.  So I renamed this to
“SETUP-SQL”.

Also, some users were using the database for other projects in addition
to their weather station, and the very generic table name “SETTINGS”
was already in use (shocking, I know).  So I also renamed the table to
“RPiWx_SETTINGS”, which is much less likely to have a naming collision.